Cleanup layout, move things to screen
